Composer unneeded

+++ source/github_hook.php
@@ -6,6 +6,7 @@ date_default_timezone_set('Europe/Paris');
 // app variables
 $app_root = realpath(__DIR__ . '/../');
 $composer = $app_root . '/composer.phar';
+$sculpin = $app_root . '/sculpin.phar';
 
 // git variables
 $branch = 'master';
@@ -25,10 +26,18 @@ function logHookResult($message, $success = false)
 }
 
 // CHECK: Download composer in the app root if it is not already there
-if (! file_exists($composer)) {
+if (! file_exists($composer) and file_exists($app_root.'/composer.json')) {
     file_put_contents(
         $composer,
         file_get_contents('https://getcomposer.org/composer.phar')
+    );
+}
+
+// Get Sculpin
+if (! file_exists($sculpin) and file_exists($app_root.'/sculpin.json')) {
+    file_put_contents(
+        $sculpin,
+        file_get_contents('https://download.sculpin.io/sculpin.phar')
     );
 }
 
@@ -44,7 +53,7 @@ if (isset($_SERVER[$header])) {
         exec("git checkout $branch ; git pull origin $branch");
 
         // Install or update dependencies
-        if (file_exists($composer)) {
+        if (file_exists($composer) and file_exists($app_root.'/composer.json')) {
             chdir($app_root);
 
             // www-data does not have a HOME or COMPOSER_HOME, create one
@@ -62,8 +71,11 @@ if (isset($_SERVER[$header])) {
         }
 
         // Generate static site
-        chdir($app_root);
-        exec('php ./vendor/sculpin/sculpin/bin/sculpin generate --env=prod > /dev/null 2>&1');
+        if (file_exists($sculpin) and file_exists($app_root.'/sculpin.json')) {
+            chdir($app_root);
+
+            exec("php {$sculpin} generate --env=prod > /dev/null 2>&1");
+        }
 
         logHookResult('Last update: ' . date('d-m-Y H:i:s'), true);
     } else {
